Reverse scrolling of inline panes
Since the action card inline panes are anchored to the bottom, reversing the scroll of the columns for all inline panes allows a more natural behavior when resizing, either due to the keyboard or other resizing sources.
By having the scroll position start at MAX_INT, this preserves the old behavior of starting at the top, instead of the bottom, which would normally be the default for a reverse scrolling column.

Fixes #2084